Model_Training/PyTorch-Lightning/requirements.txt
gdown
rarfile
lightning 
efficientnet-pytorch
comet-ml
